% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMalariaTES-package.R
\docType{package}
\name{BayesMalariaTES-package}
\alias{BayesMalariaTES}
\alias{BayesMalariaTES-package}
\title{BayesMalariaTES: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Veronica-adhiambo-ochieng/BayesMalariaTES}
  \item Report bugs at \url{https://github.com/Veronica-adhiambo-ochieng/BayesMalariaTES/issues}
}

}
\author{
\strong{Maintainer}: Veronica Adhiambo Ochieng \email{veronica.adhiambo@aims.ac.rw}

Authors:
\itemize{
  \item Mateusz Plucinski
  \item Monica Golumbeanu
}

}
\keyword{internal}
